Added the MapStruct dependency to the project to implement the Domain approach, translate the entity classes to domain classes and set the structure for the ProductRepository

+++ src/main/java/com/platzi/market/persistence/entity/Producto.java
@@ -21,7 +21,7 @@ public class Producto {
     private String nombre;
 
     @Column(name = "id_categoria")
-    private Integer idCategoria;
+    private Long idCategoria;
 
     @Column(name = "codigo_barras")
     private String codigoBarras;
@@ -55,11 +55,11 @@ public class Producto {
         this.nombre = nombre;
     }
 
-    public Integer getIdCategoria() {
+    public Long getIdCategoria() {
         return idCategoria;
     }
 
-    public void setIdCategoria(Integer idCategoria) {
+    public void setIdCategoria(Long idCategoria) {
         this.idCategoria = idCategoria;
     }
